% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_response.R
\name{client_responses}
\alias{client_responses}
\title{Function to sent HTTP request to OpenAI}
\usage{
client_responses(body_list, content)
}
\arguments{
\item{body_list}{default OpenAI parameters}

\item{content}{Parsed content}
}
\description{
Function to sent HTTP request to OpenAI
}
