% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{extract_ggplot_code}
\alias{extract_ggplot_code}
\title{Extract ggplot from qmd/rmd}
\usage{
extract_ggplot_code(file_path)
}
\arguments{
\item{file_path}{Path to the QMD/RMD file}
}
\description{
Extract ggplot from qmd/rmd
}
