% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{extract_paragraph}
\alias{extract_paragraph}
\title{Function that extract reference text using chunk_label}
\usage{
extract_paragraph(content, line_number)
}
\arguments{
\item{content}{QMD/RMD file in .txt format}

\item{line_number}{Line number that contains `@chunk_label`}
}
\description{
Function that extract reference text using chunk_label
}
