% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{find_chunk_location}
\alias{find_chunk_location}
\title{Function to find chunk location in the original document}
\usage{
find_chunk_location(content, chunk_label, code)
}
\arguments{
\item{content}{QMD/RMD file in .txt format}

\item{chunk_label}{Chunk label}

\item{code}{Source code}
}
\description{
Function to find chunk location in the original document
}
