% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{find_reference_text}
\alias{find_reference_text}
\title{Function to find reference text}
\usage{
find_reference_text(content, chunk_label, chunk_start, chunk_end)
}
\arguments{
\item{content}{QMD/RMD file in .txt format}

\item{chunk_label}{Chunk label}

\item{chunk_start}{Line number that the chunk starts}

\item{chunk_end}{Line number that the chunk ends}
}
\description{
Function to find reference text
}
