% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_response.R
\name{generate_alt_text.default}
\alias{generate_alt_text.default}
\title{Default reserved for unknown class}
\usage{
\method{generate_alt_text}{default}(input = NULL, ...)
}
\description{
Default reserved for unknown class
}
