% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_response.R
\name{lookup_captions}
\alias{lookup_captions}
\title{This function returns a list of length n where n = length(input_files)}
\usage{
lookup_captions(captions, files, combine)
}
\description{
This function returns a list of length n where n = length(input_files)
}
