% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{parse_code}
\alias{parse_code}
\title{Function to extract all code chunks from a .txt file}
\usage{
parse_code(purled_content)
}
\arguments{
\item{purled_content}{Output from knitr::purl() function}
}
\description{
Function to extract all code chunks from a .txt file
}
