% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_response.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Function to generate alt-text for data visualisations in a Quarto or R Markdown file}
\usage{
generate_alt_text(
  flnm = NULL,
  outfile = NULL,
  openai_model = "gpt-4.1",
  api = NULL,
  user_instruct = ""
)
}
\arguments{
\item{flnm}{Character string. Path and file name for the qmd or rmd file containing the plots.}

\item{outfile}{Character string. Path and file name for the output file of alt-text. If not provided will write to alt-text.txt in current folder.}

\item{openai_model}{Character string. Name of the OpenAI model used to generate alt-text.}

\item{api}{Character string. OpenAI API key used for authentication.}

\item{user_instruct}{Character string (optional). Additional user instructions to refine the style or content of the alt-text; to be appended the default system prompt.}
}
\description{
Function to generate alt-text for data visualisations in a Quarto or R Markdown file
}
