% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_response.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Function to generate alt-text}
\usage{
generate_alt_text(file_path = NULL, api = NULL, userinstruct = "")
}
\arguments{
\item{file_path}{Path to the QMD/RMD file}

\item{api}{API key}

\item{userinstruct}{Optional user instruction. This will be appended to the system prompt.}
}
\description{
Function to generate alt-text
}
